ploswater: A dataset of papers from PLOS Water journal

This dataset includes data on papers published in the PLOS Water journal. All raw data is obtained using the OpenAlex API. The data includes details on the publication, references, authorship, and institutions.

Usage

ploswater

Format

A tibble with 240 rows and 1506 variables

id

Unique Identifier

doi

DOI of the study

title

Title of the study

display_name

Displayed Name in OpenAlex

publication_year

Year of publication

publication_date

Date of publication

language

Original publication language

type

Indicates the type of scholarly work

type_crossref

The work type as specified by Crossref

indexed_in

authorships

Contains detailed information about authors and their institutional affiliations

institution_assertions

Information about institutional affiliations

countries_distinct_count

Total number of unique country codes among the work's authorships

institutions_distinct_count

Total number of unique institutions among the work's authorships

corresponding_author_ids

List of OpenAlex IDs for corresponding authors

corresponding_institution_ids

List of OpenAlex IDs for institutions of corresponding authors

fwci

Field-weighted Citation Impact, calculated as the ratio of citations received to citations expected in the publication year and three following years

has_fulltext

Boolean indicating if the full text is searchable in OpenAlex

fulltext_origin

Indicates how the full text was obtained (e.g., "pdf" or "ngrams") when has_fulltext is true

cited_by_count

Number of times the work has been cited

is_retracted

Wether study was retracted or not

is_paratext

topics

keywords

Keywords in the text

concepts

Key concepts in the text

mesh

locations_count

Number of different places where the work can be found

locations

List of all places where the work exists, including journals, proceedings, repositories (like arXiv), and institutional repositories

sustainable_development_goals

Classifications linking the work to UN Sustainable Development Goals (SDGs)

grants

Information about research funding and grants associated with the work

datasets

Associated research datasets

versions

Different versions of the work available across locations (e.g., peer-reviewed published version, preprint version)

referenced_works_count

Number of works this work cites

referenced_works

List of works cited by this work

related_works

Works that are related to this one

cited_by_api_url

URL to access the list of works citing this work

counts_by_year

Year-by-year breakdown of citation counts, excluding citations older than ten years

updated_date

ISO 8601 date string showing when the record was last modified

created_date

ISO 8601 date string indicating when the work was first added to OpenAlex

ids.openalex

The primary identifier in OpenAlex

ids.doi

The Digital Object Identifier, serving as the Canonical External ID

primary_location.is_oa

primary_location.landing_page_url

primary_location.pdf_url

primary_location.license

primary_location.license_id

primary_location.version

primary_location.is_accepted

primary_location.is_published

primary_location.source.id

primary_location.source.display_name

primary_location.source.issn_l

primary_location.source.issn

primary_location.source.is_oa

primary_location.source.is_in_doaj

primary_location.source.is_core

primary_location.source.host_organization

primary_location.source.host_organization_name

primary_location.source.host_organization_lineage

primary_location.source.host_organization_lineage_names

primary_location.source.type

open_access.is_oa

open_access.oa_status

open_access.oa_url

open_access.any_repository_has_fulltext

apc_list.value

apc_list.currency

apc_list.value_usd

apc_list.provenance

apc_paid.value

apc_paid.currency

apc_paid.value_usd

apc_paid.provenance

citation_normalized_percentile.value

citation_normalized_percentile.is_in_top_1_percent

citation_normalized_percentile.is_in_top_10_percent

cited_by_percentile_year.min

cited_by_percentile_year.max

biblio.volume

biblio.issue

biblio.first_page

biblio.last_page

primary_topic.id

primary_topic.display_name

primary_topic.score

primary_topic.subfield.id

primary_topic.subfield.display_name

primary_topic.field.id

primary_topic.field.display_name

primary_topic.domain.id

primary_topic.domain.display_name

best_oa_location.is_oa

best_oa_location.landing_page_url

best_oa_location.pdf_url

best_oa_location.license

best_oa_location.license_id

best_oa_location.version

best_oa_location.is_accepted

best_oa_location.is_published

best_oa_location.source.id

best_oa_location.source.display_name

best_oa_location.source.issn_l

best_oa_location.source.issn

best_oa_location.source.is_oa

best_oa_location.source.is_in_doaj

best_oa_location.source.is_core

best_oa_location.source.host_organization

best_oa_location.source.host_organization_name

best_oa_location.source.host_organization_lineage

best_oa_location.source.host_organization_lineage_names

best_oa_location.source.type

abstract_inverted_index.water

abstract_inverted_index.and

abstract_inverted_index.sanitation

abstract_inverted_index.programs

abstract_inverted_index.have

abstract_inverted_index.focused

abstract_inverted_index.on

abstract_inverted_index.value

abstract_inverted_index.in

abstract_inverted_index.improving

abstract_inverted_index.conditions

abstract_inverted_index.evidence

abstract_inverted_index.This

abstract_inverted_index.review

abstract_inverted_index.identified

abstract_inverted_index.empirical

abstract_inverted_index.research

abstract_inverted_index.that

abstract_inverted_index.and/or

abstract_inverted_index.reported

abstract_inverted_index.used,

abstract_inverted_index.methods

abstract_inverted_index.if

abstract_inverted_index.was

abstract_inverted_index.the

abstract_inverted_index.A

abstract_inverted_index.conceptual

abstract_inverted_index.model

abstract_inverted_index.of

abstract_inverted_index.which

abstract_inverted_index.includes

abstract_inverted_index.three

abstract_inverted_index.resources,

abstract_inverted_index.institutional

abstract_inverted_index.analysis.

abstract_inverted_index.We

abstract_inverted_index.Global

abstract_inverted_index.for

abstract_inverted_index.peer-reviewed

abstract_inverted_index.sources

abstract_inverted_index.either

abstract_inverted_index.related

abstract_inverted_index.terms

abstract_inverted_index.from

abstract_inverted_index.were

abstract_inverted_index.representing

abstract_inverted_index.over

abstract_inverted_index.used

abstract_inverted_index.followed

abstract_inverted_index.framework

abstract_inverted_index.analysis

abstract_inverted_index.approach,

abstract_inverted_index.using

abstract_inverted_index.as

abstract_inverted_index.two

abstract_inverted_index.additional

abstract_inverted_index.relevant

abstract_inverted_index.to

abstract_inverted_index.The

abstract_inverted_index.majority

abstract_inverted_index.or

abstract_inverted_index.published

abstract_inverted_index.since

abstract_inverted_index.proportion

abstract_inverted_index.studies

abstract_inverted_index.half

abstract_inverted_index.yet

abstract_inverted_index.only

abstract_inverted_index.provided

abstract_inverted_index.a

abstract_inverted_index.least

abstract_inverted_index.assessing

abstract_inverted_index.are

abstract_inverted_index.is

abstract_inverted_index.main

abstract_inverted_index.more

abstract_inverted_index.approaches

abstract_inverted_index.reduce

abstract_inverted_index.systemic

abstract_inverted_index.resources

abstract_inverted_index.For

abstract_inverted_index.first

abstract_inverted_index.time

abstract_inverted_index.has

abstract_inverted_index.been

abstract_inverted_index.focus

abstract_inverted_index.both

abstract_inverted_index.2

abstract_inverted_index.we

abstract_inverted_index.argue

abstract_inverted_index.here

abstract_inverted_index.not

abstract_inverted_index.attention

abstract_inverted_index.it

abstract_inverted_index.scientists

abstract_inverted_index.policymakers

abstract_inverted_index.several

abstract_inverted_index.historical

abstract_inverted_index.temperature

abstract_inverted_index.change

abstract_inverted_index.further

abstract_inverted_index.increased

abstract_inverted_index.with

abstract_inverted_index.use

abstract_inverted_index.global

abstract_inverted_index.levels

abstract_inverted_index.by

abstract_inverted_index.an

abstract_inverted_index.aim

abstract_inverted_index.be

abstract_inverted_index.consistent

abstract_inverted_index.current

abstract_inverted_index.policy

abstract_inverted_index.increasing

abstract_inverted_index.extreme

abstract_inverted_index.weather

abstract_inverted_index.sometimes

abstract_inverted_index.changes

abstract_inverted_index.such

abstract_inverted_index.land

abstract_inverted_index.variability

abstract_inverted_index.cycle

abstract_inverted_index.season

abstract_inverted_index.season,

abstract_inverted_index.year

abstract_inverted_index.also

abstract_inverted_index.impacts

abstract_inverted_index.all

abstract_inverted_index.needed

abstract_inverted_index.understanding

abstract_inverted_index.effectiveness

abstract_inverted_index.future

abstract_inverted_index.adaptation

abstract_inverted_index.strategies

abstract_inverted_index.reducing

abstract_inverted_index.water-related

abstract_inverted_index.climate

abstract_inverted_index.Finally,

abstract_inverted_index.role

abstract_inverted_index.assessment

abstract_inverted_index.require

abstract_inverted_index.whose

abstract_inverted_index.feasibility

abstract_inverted_index.efficiency

abstract_inverted_index.resources.

abstract_inverted_index.It

abstract_inverted_index.thus

abstract_inverted_index.essential

abstract_inverted_index.develop

abstract_inverted_index.integrated

abstract_inverted_index.approach

abstract_inverted_index.change,

abstract_inverted_index.would

abstract_inverted_index.between

abstract_inverted_index.hydrological

abstract_inverted_index.testing

abstract_inverted_index.novel

abstract_inverted_index.tool

abstract_inverted_index.public

abstract_inverted_index.health

abstract_inverted_index.monitoring,

abstract_inverted_index.but

abstract_inverted_index.laboratory

abstract_inverted_index.capacity

abstract_inverted_index.provide

abstract_inverted_index.monitoring

abstract_inverted_index.at

abstract_inverted_index.locations

abstract_inverted_index.where

abstract_inverted_index.potential

abstract_inverted_index.practices

abstract_inverted_index.wastewater

abstract_inverted_index.currently

abstract_inverted_index.quality

abstract_inverted_index.control

abstract_inverted_index.may

abstract_inverted_index.across

abstract_inverted_index.many

abstract_inverted_index.academic

abstract_inverted_index.out

abstract_inverted_index.Here,

abstract_inverted_index.our

abstract_inverted_index.including

abstract_inverted_index.standard

abstract_inverted_index.operating

abstract_inverted_index.present

abstract_inverted_index.data

abstract_inverted_index.discuss

abstract_inverted_index.reporting

abstract_inverted_index.this

abstract_inverted_index.information

abstract_inverted_index.will

abstract_inverted_index.others

abstract_inverted_index.developing

abstract_inverted_index.support

abstract_inverted_index.decision-making.

abstract_inverted_index.agriculture

abstract_inverted_index.significant

abstract_inverted_index.food

abstract_inverted_index.socio-economic

abstract_inverted_index.people

abstract_inverted_index.living

abstract_inverted_index.some

abstract_inverted_index.UN

abstract_inverted_index.Organization

abstract_inverted_index.(e.g.,

abstract_inverted_index.World

abstract_inverted_index.River

abstract_inverted_index.these

abstract_inverted_index.areas

abstract_inverted_index.resource

abstract_inverted_index.sustainability

abstract_inverted_index.its

abstract_inverted_index.unique

abstract_inverted_index.risk

abstract_inverted_index.process

abstract_inverted_index.results

abstract_inverted_index.multiple

abstract_inverted_index.natural

abstract_inverted_index.recently

abstract_inverted_index.river

abstract_inverted_index.sustainable

abstract_inverted_index.development

abstract_inverted_index.particularly

abstract_inverted_index.lack

abstract_inverted_index.leads

abstract_inverted_index.discharge

abstract_inverted_index.flow

abstract_inverted_index.agricultural

abstract_inverted_index.plant

abstract_inverted_index.uptake

abstract_inverted_index.effects

abstract_inverted_index.crop

abstract_inverted_index.growth

abstract_inverted_index.higher

abstract_inverted_index.lower

abstract_inverted_index.soil

abstract_inverted_index.microbial

abstract_inverted_index.activity,

abstract_inverted_index.overall

abstract_inverted_index.case

abstract_inverted_index.critical

abstract_inverted_index.regions

abstract_inverted_index.can

abstract_inverted_index.considered

abstract_inverted_index.possible

abstract_inverted_index.Two

abstract_inverted_index.among

abstract_inverted_index.home

abstract_inverted_index.animal

abstract_inverted_index.strong

abstract_inverted_index.rural

abstract_inverted_index.major

abstract_inverted_index.area

abstract_inverted_index.important

abstract_inverted_index.hydraulic

abstract_inverted_index.works

abstract_inverted_index.Republic

abstract_inverted_index.region

abstract_inverted_index.faces

abstract_inverted_index.due

abstract_inverted_index4,

abstract_inverted_index.increasingly

abstract_inverted_index.challenging

abstract_inverted_index.recent

abstract_inverted_index.attributed

abstract_inverted_index.range

abstract_inverted_index.building

abstract_inverted_index.studied

abstract_inverted_index.February

abstract_inverted_index.biological

abstract_inverted_index.parameters

abstract_inverted_index.pH,

abstract_inverted_index.total

abstract_inverted_index.count

abstract_inverted_index.analyzed

abstract_inverted_index.understand

abstract_inverted_index.differences

abstract_inverted_index.within

abstract_inverted_index.chlorine

abstract_inverted_index.concentration

abstract_inverted_index.decreased

abstract_inverted_index.(

abstract_inverted_index.p

abstract_inverted_index.<

abstract_inverted_index.point

abstract_inverted_index.plumbing

abstract_inverted_index.As

abstract_inverted_index.much

abstract_inverted_index.greater

abstract_inverted_index.compared

abstract_inverted_index.every

abstract_inverted_index.sampling

abstract_inverted_index.found

abstract_inverted_index.highest

abstract_inverted_index.no

abstract_inverted_index.recorded

abstract_inverted_index.during

abstract_inverted_index.lead

abstract_inverted_index.health-based

abstract_inverted_index.drinking

abstract_inverted_index.four

abstract_inverted_index.same

abstract_inverted_index.shared

abstract_inverted_index.use.

abstract_inverted_index.conducted

abstract_inverted_index.one

abstract_inverted_index.required

abstract_inverted_index.supply.

abstract_inverted_index.viable

abstract_inverted_index.method

abstract_inverted_index.need

abstract_inverted_index.19

abstract_inverted_index.than

abstract_inverted_index.50

abstract_inverted_index.before

abstract_inverted_index.prior

abstract_inverted_index.suggest

abstract_inverted_index.significantly

abstract_inverted_index.implications

abstract_inverted_index.collected

abstract_inverted_index.infrastructure

abstract_inverted_index.engineering

abstract_inverted_index.security

abstract_inverted_index.In

abstract_inverted_index.freshwater

abstract_inverted_index.society

abstract_inverted_index.environment

abstract_inverted_index.could

abstract_inverted_index.benefit

abstract_inverted_index.Such

abstract_inverted_index.conservation

abstract_inverted_index.ongoing

abstract_inverted_index.species

abstract_inverted_index.their

abstract_inverted_index.systems.

abstract_inverted_index.However,

abstract_inverted_index.make

abstract_inverted_index.meaningful

abstract_inverted_index.contributions

abstract_inverted_index.must

abstract_inverted_index.established

abstract_inverted_index.link

abstract_inverted_index.common

abstract_inverted_index.six

abstract_inverted_index.actions

abstract_inverted_index.life

abstract_inverted_index.science

abstract_inverted_index.highlight

abstract_inverted_index.knowledge

abstract_inverted_index.necessary

abstract_inverted_index.how

abstract_inverted_index.play

abstract_inverted_index.direct

abstract_inverted_index.aquatic

abstract_inverted_index.quality,

abstract_inverted_index.ecosystems.

abstract_inverted_index.ensure

abstract_inverted_index.ecosystems

abstract_inverted_index.species.

abstract_inverted_index.Water

abstract_inverted_index.managers

abstract_inverted_index.consider

abstract_inverted_index.whether

abstract_inverted_index.concern

abstract_inverted_index.reservoir

abstract_inverted_index.management,

abstract_inverted_index.modeling

abstract_inverted_index.requires

abstract_inverted_index.models

abstract_inverted_index.watershed

abstract_inverted_index.various

abstract_inverted_index.most

abstract_inverted_index.practical

abstract_inverted_index.selecting

abstract_inverted_index.availability

abstract_inverted_index.number

abstract_inverted_index.decisions

abstract_inverted_index.regarding

abstract_inverted_index.best

abstract_inverted_index.relative

abstract_inverted_index.importance

abstract_inverted_index.different

abstract_inverted_index.limited

abstract_inverted_index.should

abstract_inverted_index.assess

abstract_inverted_index.existing

abstract_inverted_index.then

abstract_inverted_index.limitations

abstract_inverted_index.questions

abstract_inverted_index.previous

abstract_inverted_index.While

abstract_inverted_index.there

abstract_inverted_index.work

abstract_inverted_index.options

abstract_inverted_index.activity

abstract_inverted_index.leading

abstract_inverted_index.widespread

abstract_inverted_index.lake

abstract_inverted_index.key

abstract_inverted_index.health.

abstract_inverted_index.anthropogenic

abstract_inverted_index.affecting

abstract_inverted_index.space

abstract_inverted_index.potentially

abstract_inverted_index.highly

abstract_inverted_index.responses

abstract_inverted_index.poorly

abstract_inverted_index.million

abstract_inverted_index.estimates

abstract_inverted_index.4

abstract_inverted_index.km

abstract_inverted_index.quantify

abstract_inverted_index.2020

abstract_inverted_index.large

abstract_inverted_index.cumulative

abstract_inverted_index.increase

abstract_inverted_index.human

abstract_inverted_index.lakes

abstract_inverted_index.positive

abstract_inverted_index.trends

abstract_inverted_index.times

abstract_inverted_index.demonstrates

abstract_inverted_index.inform

abstract_inverted_index.management

abstract_inverted_index.determine

abstract_inverted_index.local

abstract_inverted_index.Existing

abstract_inverted_index.literature

abstract_inverted_index.water,

abstract_inverted_index.sanitation,

abstract_inverted_index.hygiene

abstract_inverted_index.(WASH)

abstract_inverted_index.United

abstract_inverted_index.States,

abstract_inverted_index.issues

abstract_inverted_index.areas.

abstract_inverted_index.emerging

abstract_inverted_index.infectious

abstract_inverted_index.disease

abstract_inverted_index.face

abstract_inverted_index.challenges

abstract_inverted_index.urban

abstract_inverted_index.study

abstract_inverted_index.WASH

abstract_inverted_index.needs

abstract_inverted_index.through

abstract_inverted_index.interviews

abstract_inverted_index.(n

abstract_inverted_index.=

abstract_inverted_index.participants

abstract_inverted_index.18

abstract_inverted_index.years

abstract_inverted_index.lived

abstract_inverted_index.five

abstract_inverted_index.had

abstract_inverted_index.Using

abstract_inverted_index.analysis,

abstract_inverted_index.factors

abstract_inverted_index.access,

abstract_inverted_index.outcomes

abstract_inverted_index.result

abstract_inverted_index.explore

abstract_inverted_index.experiences

abstract_inverted_index.who

abstract_inverted_index.those

abstract_inverted_index.did

abstract_inverted_index.not.

abstract_inverted_index.Our

abstract_inverted_index.findings

abstract_inverted_index.revealed

abstract_inverted_index.contributed

abstract_inverted_index.physical,

abstract_inverted_index.relationship

abstract_inverted_index.communities

abstract_inverted_index.impacted

abstract_inverted_index.facilities

abstract_inverted_index.safe

abstract_inverted_index.available

abstract_inverted_index.inadequate

abstract_inverted_index.access

abstract_inverted_index.collaboration

abstract_inverted_index.reduction

abstract_inverted_index.services

abstract_inverted_index.example,

abstract_inverted_index.hand

abstract_inverted_index.community

abstract_inverted_index.measures

abstract_inverted_index.transmission

abstract_inverted_index.COVID-19

abstract_inverted_index.other

abstract_inverted_index.diseases.

abstract_inverted_index.guidance

abstract_inverted_index.standards

abstract_inverted_index.new

abstract_inverted_index.tools

abstract_inverted_index.targeted

abstract_inverted_index.assessments

abstract_inverted_index.pandemic.

abstract_inverted_index.organizations

abstract_inverted_index.evaluate

abstract_inverted_index.interventions

abstract_inverted_index.mitigate

abstract_inverted_index.middle-income

abstract_inverted_index.countries

abstract_inverted_index.focusing

abstract_inverted_index.household

abstract_inverted_index.settings

abstract_inverted_index.observed

abstract_inverted_index.gaps

abstract_inverted_index.specifically

abstract_inverted_index.often

abstract_inverted_index.about

abstract_inverted_index.high

abstract_inverted_index.after

abstract_inverted_index.stations

abstract_inverted_index.points

abstract_inverted_index.external

abstract_inverted_index.needed.

abstract_inverted_index.refinement

abstract_inverted_index.help

abstract_inverted_index.aspects

abstract_inverted_index.efforts

abstract_inverted_index.remain

abstract_inverted_index.functional

abstract_inverted_index.inclusive

abstract_inverted_index.U.S.

abstract_inverted_index.service

abstract_inverted_index.health,

abstract_inverted_index.economic

abstract_inverted_index.expressed

abstract_inverted_index.sample

abstract_inverted_index.approximately

abstract_inverted_index.city

abstract_inverted_index.perceptions

abstract_inverted_index.barriers

abstract_inverted_index.meeting

abstract_inverted_index.question

abstract_inverted_index.cost

abstract_inverted_index.capital

abstract_inverted_index.projects

abstract_inverted_index.increases

abstract_inverted_index.infrastructure.

abstract_inverted_index.demonstrate

abstract_inverted_index.disparities

abstract_inverted_index.great

abstract_inverted_index.little

abstract_inverted_index.condition

abstract_inverted_index.making

abstract_inverted_index.financial

abstract_inverted_index.decision

abstract_inverted_index.clean

abstract_inverted_index.substantial

abstract_inverted_index.States

abstract_inverted_index.cannot

abstract_inverted_index.pathways

abstract_inverted_index.influence

abstract_inverted_index.behaviors

abstract_inverted_index.vulnerable

abstract_inverted_index.interviewed

abstract_inverted_index.low-income

abstract_inverted_index.bill

abstract_inverted_index.content

abstract_inverted_index.interview

abstract_inverted_index.Results

abstract_inverted_index.showed

abstract_inverted_index.ways

abstract_inverted_index.made

abstract_inverted_index.times.

abstract_inverted_index.influenced

abstract_inverted_index.utility

abstract_inverted_index.included

abstract_inverted_index.reduced

abstract_inverted_index.decline

abstract_inverted_index.government

abstract_inverted_index.affordable

abstract_inverted_index.households.

abstract_inverted_index.infrastructure,

abstract_inverted_index.state

abstract_inverted_index.well

abstract_inverted_index.municipal

abstract_inverted_index.policies

abstract_inverted_index.households

abstract_inverted_index.growing

abstract_inverted_index.residents

abstract_inverted_index.balance

abstract_inverted_index.resulting

abstract_inverted_index.affect

abstract_inverted_index.loss

abstract_inverted_index.impact

abstract_inverted_index.system,

abstract_inverted_index.influencing

abstract_inverted_index.frequency

abstract_inverted_index.During

abstract_inverted_index.past

abstract_inverted_index.basin

abstract_inverted_index.uses

abstract_inverted_index.along

abstract_inverted_index.predicted

abstract_inverted_index.temporal

abstract_inverted_index.spatial

abstract_inverted_index.sediment

abstract_inverted_index.effective

abstract_inverted_index.50%

abstract_inverted_index.baseline

abstract_inverted_index.cause

abstract_inverted_index.3%

abstract_inverted_index.mean

abstract_inverted_index.surface

abstract_inverted_index.solid

abstract_inverted_index.estimated

abstract_inverted_index.concentrations

abstract_inverted_index.cover

abstract_inverted_index.biochar

abstract_inverted_index.remove

abstract_inverted_index.improve

abstract_inverted_index.P

abstract_inverted_index.molecular

abstract_inverted_index.carbon

abstract_inverted_index.produced

abstract_inverted_index.under

abstract_inverted_index.3

abstract_inverted_index.gas

abstract_inverted_index.mass

abstract_inverted_index.capacities

abstract_inverted_index.near

abstract_inverted_index.structure

abstract_inverted_index.indicated

abstract_inverted_index.advance

abstract_inverted_index.treatment

abstract_inverted_index.processes

abstract_inverted_index.employed

abstract_inverted_index.stormwater

abstract_inverted_index.stream

abstract_inverted_index.performance

abstract_inverted_index.metrics

abstract_inverted_index.they

abstract_inverted_index.catchment

abstract_inverted_index.aimed

abstract_inverted_index.response

abstract_inverted_index.application

abstract_inverted_index.experiment

abstract_inverted_index.test

abstract_inverted_index.improved

abstract_inverted_index.achieved

abstract_inverted_index.experimental

abstract_inverted_index.similar

abstract_inverted_index.per

abstract_inverted_index.implementation

abstract_inverted_index.achieve

abstract_inverted_index.success

abstract_inverted_index.final

abstract_inverted_index.into

abstract_inverted_index.demand

abstract_inverted_index.measure

abstract_inverted_index.period

abstract_inverted_index.varied

abstract_inverted_index.modelling

abstract_inverted_index.broader

abstract_inverted_index.decades

abstract_inverted_index.environmental

abstract_inverted_index.surveillance

abstract_inverted_index.pathogens

abstract_inverted_index.single

abstract_inverted_index.pathogen

abstract_inverted_index.small

abstract_inverted_index.few

abstract_inverted_index.quantitative

abstract_inverted_index.developed

abstract_inverted_index.targets

abstract_inverted_index.applied

abstract_inverted_index.samples

abstract_inverted_index.plants

abstract_inverted_index.From

abstract_inverted_index.detected

abstract_inverted_index.find

abstract_inverted_index.E

abstract_inverted_index..

abstract_inverted_index.coli

abstract_inverted_index.stable

abstract_inverted_index.notable

abstract_inverted_index.,

abstract_inverted_index.broad

abstract_inverted_index.expanding

abstract_inverted_index.fecal

abstract_inverted_index.waste

abstract_inverted_index.solids

abstract_inverted_index.changed

abstract_inverted_index.objective

abstract_inverted_index.One

abstract_inverted_index.when

abstract_inverted_index.To

abstract_inverted_index.trend

abstract_inverted_index.index

abstract_inverted_index.measurements

abstract_inverted_index.characterize

abstract_inverted_index.trends.

abstract_inverted_index.better

abstract_inverted_index.so

abstract_inverted_index.5

abstract_inverted_index.detect

abstract_inverted_index.respectively.

abstract_inverted_index.providing

abstract_inverted_index.changing

abstract_inverted_index.2021

abstract_inverted_index.without

abstract_inverted_index.domestic

abstract_inverted_index.systems

abstract_inverted_index.performed

abstract_inverted_index.combined

abstract_inverted_index.Texas

abstract_inverted_index.affected

abstract_inverted_index.populations

abstract_inverted_index.cross-sectional

abstract_inverted_index.survey

abstract_inverted_index.shows

abstract_inverted_index.size

abstract_inverted_index.system.

abstract_inverted_index.mostly

abstract_inverted_index.disproportionately

abstract_inverted_index.likely

abstract_inverted_index.utilities

abstract_inverted_index.respondents

abstract_inverted_index.resilience

abstract_inverted_index.efficient

abstract_inverted_index.emergency

abstract_inverted_index.physical

abstract_inverted_index.based

abstract_inverted_index.system

abstract_inverted_index.complex

abstract_inverted_index.conventional

abstract_inverted_index.remote

abstract_inverted_index.sensing

abstract_inverted_index.New

abstract_inverted_index.addressing

abstract_inverted_index.mainly

abstract_inverted_index.types

abstract_inverted_index.problems

abstract_inverted_index.challenges,

abstract_inverted_index.identify

abstract_inverted_index.opportunities

abstract_inverted_index.hydrology

abstract_inverted_index.around

abstract_inverted_index.insights

abstract_inverted_index.just

abstract_inverted_index.propose

abstract_inverted_index.larger

abstract_inverted_index.big

abstract_inverted_index.call

abstract_inverted_index.action

abstract_inverted_index.context

abstract_inverted_index.practice

abstract_inverted_index.sector

abstract_inverted_index.private

abstract_inverted_index.engagement

abstract_inverted_index.good

abstract_inverted_index.very

abstract_inverted_index.issue

abstract_inverted_index.pollution

abstract_inverted_index.throughout

abstract_inverted_index.entire

abstract_inverted_index.West

abstract_inverted_index.distribution

abstract_inverted_index.pollutants

abstract_inverted_index.considering

abstract_inverted_index.level,

abstract_inverted_index.next

abstract_inverted_index.generated

abstract_inverted_index.level

abstract_inverted_index.each

abstract_inverted_index.selected

abstract_inverted_index.estimate

abstract_inverted_index.low

abstract_inverted_index.alternative

abstract_inverted_index.solution

abstract_inverted_index.globally

abstract_inverted_index.precipitation,

abstract_inverted_index.processes.

abstract_inverted_index.performance.

abstract_inverted_index.estimating

abstract_inverted_index.characteristics

abstract_inverted_index.remains

abstract_inverted_index.models,

abstract_inverted_index.method.

abstract_inverted_index.Spatial

abstract_inverted_index.associated

abstract_inverted_index.cold

abstract_inverted_index.association

abstract_inverted_index.variations

abstract_inverted_index.operational

abstract_inverted_index.full-scale

abstract_inverted_index.integration

abstract_inverted_index.chemical

abstract_inverted_index.accumulation

abstract_inverted_index.being

abstract_inverted_index.driver

abstract_inverted_index.upon

abstract_inverted_index.complex,

abstract_inverted_index.indicating

abstract_inverted_index.These

abstract_inverted_index.water.

abstract_inverted_index.By

abstract_inverted_index.process,

abstract_inverted_index.targeting

abstract_inverted_index.long-term

abstract_inverted_index.quality.

abstract_inverted_index.urbanization,

abstract_inverted_index.population

abstract_inverted_index.inequality

abstract_inverted_index.study,

abstract_inverted_index.peri-urban

abstract_inverted_index.examine

abstract_inverted_index.explained

abstract_inverted_index.regression

abstract_inverted_index.source

abstract_inverted_index.pressure

abstract_inverted_index.supply

abstract_inverted_index.odds

abstract_inverted_index.ratio

abstract_inverted_index.factor

abstract_inverted_index.Findings

abstract_inverted_index.even

abstract_inverted_index.services,

abstract_inverted_index.stakeholders

abstract_inverted_index.guide

abstract_inverted_index.infrastructural

abstract_inverted_index.settings.

abstract_inverted_index.dynamics

abstract_inverted_index.transboundary

abstract_inverted_index.nature

abstract_inverted_index.cross

abstract_inverted_index.rivers,

abstract_inverted_index.Thus,

abstract_inverted_index.defined

abstract_inverted_index.beyond

abstract_inverted_index.field

abstract_inverted_index.concept.

abstract_inverted_index.With

abstract_inverted_index.riverine

abstract_inverted_index.become

abstract_inverted_index.crucial

abstract_inverted_index.study.

abstract_inverted_index.i.e.,

abstract_inverted_index.Yet,

abstract_inverted_index.primary

abstract_inverted_index.because

abstract_inverted_index.cities

abstract_inverted_index.Moreover,

abstract_inverted_index.what

abstract_inverted_index.them

abstract_inverted_index.do

abstract_inverted_index.ability

abstract_inverted_index.extensive

abstract_inverted_index.qualitative

abstract_inverted_index.South

abstract_inverted_index.purpose

abstract_inverted_index.article

abstract_inverted_index.poor

abstract_inverted_index.toward

abstract_inverted_index.community.

abstract_inverted_index.include

abstract_inverted_index.education

abstract_inverted_index.towards

abstract_inverted_index.equitable

abstract_inverted_index.and,

abstract_inverted_index.on-site

abstract_inverted_index.aims

abstract_inverted_index.taken

abstract_inverted_index.parameters.

abstract_inverted_index.hydrologic

abstract_inverted_index.while

abstract_inverted_index.assessed

abstract_inverted_index.downstream

abstract_inverted_index.difficulties

abstract_inverted_index.simulate

abstract_inverted_index.conditions,

abstract_inverted_index.contaminants.

abstract_inverted_index.scale

abstract_inverted_index.National

abstract_inverted_index.Monitoring

abstract_inverted_index.monitored

abstract_inverted_index.antimicrobial

abstract_inverted_index.alignment

abstract_inverted_index.Health

abstract_inverted_index.planning,

abstract_inverted_index.waters

abstract_inverted_index.function

abstract_inverted_index.humans,

abstract_inverted_index.animals,

abstract_inverted_index.environment.

abstract_inverted_index.sequencing

abstract_inverted_index.coverage

abstract_inverted_index.shown

abstract_inverted_index.enhance

abstract_inverted_index.report

abstract_inverted_index.profile

abstract_inverted_index.distinct

abstract_inverted_index.sites

abstract_inverted_index.enriched

abstract_inverted_index.almost

abstract_inverted_index.30%

abstract_inverted_index.Additionally,

abstract_inverted_index.specific

abstract_inverted_index.sites.

abstract_inverted_index.Development

abstract_inverted_index.technologies

abstract_inverted_index.vulnerability

abstract_inverted_index.scheme

abstract_inverted_index.design

abstract_inverted_index.period.

abstract_inverted_index.variables

abstract_inverted_index.less

abstract_inverted_index.treatment.

abstract_inverted_index.conclude

abstract_inverted_index.appropriate

abstract_inverted_index.urgent

abstract_inverted_index.set

abstract_inverted_index.pressures

abstract_inverted_index.wastewater,

abstract_inverted_index.maintain

abstract_inverted_index.involve

abstract_inverted_index.general

abstract_inverted_index.way

abstract_inverted_index.get

abstract_inverted_index.participation

abstract_inverted_index.world

abstract_inverted_index.take

abstract_inverted_index.groundwater

abstract_inverted_index.science,

abstract_inverted_index.respect

abstract_inverted_index.authorities

abstract_inverted_index.social

abstract_inverted_index.knowledge,

abstract_inverted_index.environment,

abstract_inverted_index.encourage

abstract_inverted_index.tap

abstract_inverted_index.effect

abstract_inverted_index.composition

abstract_inverted_index.relatively

abstract_inverted_index.simple

abstract_inverted_index.adapted

abstract_inverted_index.expertise

abstract_inverted_index.manage

abstract_inverted_index.safety

abstract_inverted_index.exacerbated

abstract_inverted_index.working

abstract_inverted_index.capacity,

abstract_inverted_index.informant

abstract_inverted_index.10

abstract_inverted_index.risks

abstract_inverted_index.led

abstract_inverted_index.adequate

abstract_inverted_index.diseases

abstract_inverted_index.investigated

abstract_inverted_index.recommended

abstract_inverted_index.active

abstract_inverted_index.training

abstract_inverted_index.flood

abstract_inverted_index.treated

abstract_inverted_index.industrial

abstract_inverted_index.demands

abstract_inverted_index.up

abstract_inverted_index.put

abstract_inverted_index.governance

abstract_inverted_index.respond

abstract_inverted_index.values

abstract_inverted_index.complexities

abstract_inverted_index.status

abstract_inverted_index.time,

abstract_inverted_index.exist

abstract_inverted_index.groups

abstract_inverted_index.regional

abstract_inverted_index.networks

abstract_inverted_index.representation

abstract_inverted_index.operate

abstract_inverted_index.activities

abstract_inverted_index.project

abstract_inverted_index.designs

abstract_inverted_index.largest

abstract_inverted_index.located

abstract_inverted_index.m

abstract_inverted_index.meet

abstract_inverted_index.income

abstract_inverted_index.census

abstract_inverted_index.users

abstract_inverted_index.).

abstract_inverted_index.rather

abstract_inverted_index.patterns,

abstract_inverted_index.patterns

abstract_inverted_index.products

abstract_inverted_index.techniques

abstract_inverted_index.addition

abstract_inverted_index.evaluated

abstract_inverted_index.monthly

abstract_inverted_index.statistical

abstract_inverted_index.indicators

abstract_inverted_index.correlation

abstract_inverted_index.consistently

abstract_inverted_index.Overall,

abstract_inverted_index.country

abstract_inverted_index.type

abstract_inverted_index.time.

abstract_inverted_index.technical

abstract_inverted_index.paper

abstract_inverted_index.researchers

abstract_inverted_index.demonstrated

abstract_inverted_index.irrigated

abstract_inverted_index.Specifically,

abstract_inverted_index.nine

abstract_inverted_index.30

abstract_inverted_index.tested

abstract_inverted_index.individually

abstract_inverted_index.obtained

abstract_inverted_index.years,

abstract_inverted_index.applicable

abstract_inverted_index.oxygen

abstract_inverted_index.effluent

abstract_inverted_index.secondary

abstract_inverted_index.costs

abstract_inverted_index.still

abstract_inverted_index.promising

abstract_inverted_index.reports

abstract_inverted_index.)

abstract_inverted_index.installed

abstract_inverted_index.months

abstract_inverted_index.measured

abstract_inverted_index.evaluation

abstract_inverted_index.show

abstract_inverted_index.nearly

abstract_inverted_index.desired

abstract_inverted_index.target

abstract_inverted_index.varying

abstract_inverted_index.measurement

abstract_inverted_index.share

abstract_inverted_index.There

abstract_inverted_index.waters,

abstract_inverted_index.detection

abstract_inverted_index.quantified

abstract_inverted_index.timely

abstract_inverted_index.waters.

abstract_inverted_index.largely

abstract_inverted_index.governance,

abstract_inverted_index.management.

abstract_inverted_index.learn

abstract_inverted_index.enabling

abstract_inverted_index.Kenya.

abstract_inverted_index.below

abstract_inverted_index.Furthermore,

abstract_inverted_index.Escherichia

abstract_inverted_index.samples,

abstract_inverted_index.3)

abstract_inverted_index.detected.

abstract_inverted_index.They

abstract_inverted_index.together

abstract_inverted_index.driven

abstract_inverted_index.rapid

abstract_inverted_index.calls

abstract_inverted_index.efforts.

abstract_inverted_index.reliable

abstract_inverted_index.international

abstract_inverted_index.national

abstract_inverted_index.improvements

abstract_inverted_index.underlying

abstract_inverted_index.map

abstract_inverted_index.implemented

abstract_inverted_index.restoration

abstract_inverted_index.interconnected

abstract_inverted_index.directly

abstract_inverted_index.opportunity

abstract_inverted_index.driving

abstract_inverted_index.vision

abstract_inverted_index.free

abstract_inverted_index.known

abstract_inverted_index.parts

abstract_inverted_index.temperature,

abstract_inverted_index.All

abstract_inverted_index.residence

abstract_inverted_index.culture

abstract_inverted_index.aeruginosa

abstract_inverted_index.storage

abstract_inverted_index.bacteria

abstract_inverted_index.presence

abstract_inverted_index.genome

abstract_inverted_index.Despite

abstract_inverted_index.pit

abstract_inverted_index.latrines

abstract_inverted_index.understood.

abstract_inverted_index.characterized

abstract_inverted_index.sludge

abstract_inverted_index.surveys

abstract_inverted_index.contain

abstract_inverted_index.area.

abstract_inverted_index.degradation

abstract_inverted_index.maintenance

abstract_inverted_index.governments

abstract_inverted_index.released

abstract_inverted_index.promote

abstract_inverted_index.actors

abstract_inverted_index.assigned

abstract_inverted_index.93

abstract_inverted_index.influences

abstract_inverted_index.formal

abstract_inverted_index.align

abstract_inverted_index.recommend

abstract_inverted_index.expand

abstract_inverted_index.shift

abstract_inverted_index.Water,

abstract_inverted_index.attributable

abstract_inverted_index.operation

abstract_inverted_index.trained

abstract_inverted_index.staff

abstract_inverted_index.sustain

abstract_inverted_index.guidelines

abstract_inverted_index.accountability

abstract_inverted_index.order

abstract_inverted_index.enhanced

abstract_inverted_index.zero

abstract_inverted_index.operator

abstract_inverted_index.reaching

abstract_inverted_index.sectors

abstract_inverted_index.relationships

abstract_inverted_index.formation

abstract_inverted_index.clustering

abstract_inverted_index.contrast,

abstract_inverted_index.Multiple

abstract_inverted_index.compounds

abstract_inverted_index.containing

abstract_inverted_index.Further

abstract_inverted_index.investigation

abstract_inverted_index.late

abstract_inverted_index.summer

abstract_inverted_index.provides

abstract_inverted_index.exposure

abstract_inverted_index.growth,

abstract_inverted_index.supply,

abstract_inverted_index.ensuring

abstract_inverted_index.quantity

abstract_inverted_index.document

abstract_inverted_index.primarily

abstract_inverted_index.Environmental

abstract_inverted_index.First,

abstract_inverted_index.organic

abstract_inverted_index.matter

abstract_inverted_index.sewage

abstract_inverted_index.operations.

abstract_inverted_index.When

abstract_inverted_index.civil

abstract_inverted_index.linked

abstract_inverted_index.examining

abstract_inverted_index.contribute

abstract_inverted_index.predictors

abstract_inverted_index.sanitation.

abstract_inverted_index.Based

abstract_inverted_index.research.

abstract_inverted_index.Ghana

abstract_inverted_index.prevalence

abstract_inverted_index.piped

abstract_inverted_index.third

abstract_inverted_index.open

abstract_inverted_index.members

abstract_inverted_index.Key

abstract_inverted_index.themes

abstract_inverted_index.1)

abstract_inverted_index.2)

abstract_inverted_index.city.

abstract_inverted_index.loads

abstract_inverted_index.Bay

abstract_inverted_index.60%

abstract_inverted_index.shorter

abstract_inverted_index.median

abstract_inverted_index.does

abstract_inverted_index.detectable

abstract_inverted_index.contamination

abstract_inverted_index.infiltration

abstract_inverted_index.recharge

abstract_inverted_index.groundwater,

abstract_inverted_index.makes

abstract_inverted_index.conduct

abstract_inverted_index.upstream

abstract_inverted_index.threats

abstract_inverted_index.activities,

abstract_inverted_index.following

abstract_inverted_index.uncertainty

abstract_inverted_index.urbanized

abstract_inverted_index.air

abstract_inverted_index.whereas

abstract_inverted_index.constant

abstract_inverted_index.years.

abstract_inverted_index.useful

abstract_inverted_index.conditions.

abstract_inverted_index.practitioners

abstract_inverted_index.rely

abstract_inverted_index.rainwater

abstract_inverted_index.presents

abstract_inverted_index.collection

abstract_inverted_index.intervention

abstract_inverted_index.completed

abstract_inverted_index.11%

abstract_inverted_index.remained

abstract_inverted_index.provision

abstract_inverted_index.locally

abstract_inverted_index.countries.

abstract_inverted_index.second

abstract_inverted_index.site

abstract_inverted_index.explored

abstract_inverted_index.interactive

abstract_inverted_index.intensity

abstract_inverted_index.Understanding

abstract_inverted_index.subsequent

abstract_inverted_index.flows

abstract_inverted_index.detailed

abstract_inverted_index.watersheds

abstract_inverted_index.Upper

abstract_inverted_index.dissolved

abstract_inverted_index.Colorado

abstract_inverted_index.previously

abstract_inverted_index.protection

abstract_inverted_index.maximum

abstract_inverted_index.contaminant

abstract_inverted_index.participatory

abstract_inverted_index.gender

abstract_inverted_index.Cambodia

abstract_inverted_index.program

abstract_inverted_index.reflect

abstract_inverted_index.extent

abstract_inverted_index.indicate

abstract_inverted_index.engaging

abstract_inverted_index.supporting

abstract_inverted_index.derived

abstract_inverted_index.sector,

abstract_inverted_index.investigate

abstract_inverted_index.adoption

abstract_inverted_index.aged

abstract_inverted_index.children

abstract_inverted_index.facilities.

abstract_inverted_index.Developing

abstract_inverted_index.journal

abstract_inverted_index.articles

abstract_inverted_index.last

abstract_inverted_index.height

abstract_inverted_index.worldwide

abstract_inverted_index.offers

abstract_inverted_index.dynamic

abstract_inverted_index.areas,

abstract_inverted_index.magnitude

abstract_inverted_index.widely

abstract_inverted_index.humanitarian

abstract_inverted_index.Research

abstract_inverted_index.online

abstract_inverted_index.score

abstract_inverted_index.according

abstract_inverted_index.experts

abstract_inverted_index.evaluating

abstract_inverted_index.contexts.

abstract_inverted_index.priorities.

abstract_inverted_index.21

abstract_inverted_index.means

abstract_inverted_index.parameters,

abstract_inverted_index.interest

abstract_inverted_index.insecurity.

abstract_inverted_index.semi-arid

abstract_inverted_index.insecurity

abstract_inverted_index.dry

abstract_inverted_index.techniques.

abstract_inverted_index.Given

abstract_inverted_index.offer

abstract_inverted_index.prioritize

abstract_inverted_index.risks,

abstract_inverted_index.filters

abstract_inverted_index.incidence

abstract_inverted_index.designed

abstract_inverted_index.risk.

abstract_inverted_index.Local

abstract_inverted_index.coordination

abstract_inverted_index.use,

abstract_inverted_index.isotope

abstract_inverted_index.(i.e.,

abstract_inverted_index.H

abstract_inverted_index.seasons

abstract_inverted_index.validated

abstract_inverted_index.annually

abstract_inverted_index.wells

abstract_inverted_index.reliability

abstract_inverted_index.overcome

abstract_inverted_index.households,

abstract_inverted_index.women

abstract_inverted_index.safely

abstract_inverted_index.managed

abstract_inverted_index.Although

abstract_inverted_index.establishing

abstract_inverted_index.divided

abstract_inverted_index.±

abstract_inverted_index.12

abstract_inverted_index.9

abstract_inverted_index.15

abstract_inverted_index.winter

abstract_inverted_index.farmers

abstract_inverted_index.actual

abstract_inverted_index.operated

abstract_inverted_index.northern

abstract_inverted_index.arsenic

abstract_inverted_index.presented

abstract_inverted_index.ranged

abstract_inverted_index.limits

abstract_inverted_index.60

abstract_inverted_index.sampled

abstract_inverted_index.former

abstract_inverted_index.wellbeing.

abstract_inverted_index.year,

abstract_inverted_index.dataset

abstract_inverted_index.suggesting

abstract_inverted_index.fully

abstract_inverted_index.San

abstract_inverted_index.Out

abstract_inverted_index.disposal

abstract_inverted_index.waterborne

abstract_inverted_index.(e.g.

abstract_inverted_index.practices,

abstract_inverted_index.volume

abstract_inverted_index.passive

abstract_inverted_index.coupled

abstract_inverted_index.faecal

abstract_inverted_index.indicator

abstract_inverted_index.contexts

abstract_inverted_index.real-time

abstract_inverted_index.achieving

abstract_inverted_index.undertaken

abstract_inverted_index.First

abstract_inverted_index.approaches,

abstract_inverted_index.planetary

abstract_inverted_index.salinity

abstract_inverted_index.resulted

abstract_inverted_index.16S

abstract_inverted_index.rRNA

abstract_inverted_index.gene

abstract_inverted_index.prioritized

abstract_inverted_index.levels.

abstract_inverted_index.prevent

abstract_inverted_index.fish

abstract_inverted_index.site.

abstract_inverted_index.removal

abstract_inverted_index.pH

abstract_inverted_index.confirmed

abstract_inverted_index.possibly

abstract_inverted_index.leaves

abstract_inverted_index.September

abstract_inverted_index.socially

abstract_inverted_index.consumption

abstract_inverted_index.linear

abstract_inverted_index.explain

abstract_inverted_index.until

abstract_inverted_index.outcomes,

abstract_inverted_index.required.

abstract_inverted_index.regulatory

abstract_inverted_index.adopted

abstract_inverted_index.connections

abstract_inverted_index.occur

abstract_inverted_index.short-term

abstract_inverted_index.wells.

abstract_inverted_index.(including

abstract_inverted_index.Both

abstract_inverted_index.events

abstract_inverted_index.contaminated

abstract_inverted_index.worked

abstract_inverted_index.knowledges

abstract_inverted_index.collaborative

abstract_inverted_index.central

abstract_inverted_index.climatic

abstract_inverted_index.outcome

abstract_inverted_index.settings,

abstract_inverted_index.extract

abstract_inverted_index.production,

abstract_inverted_index.comparison

abstract_inverted_index.tank

abstract_inverted_index.log

abstract_inverted_index.samples.

abstract_inverted_index.Therefore,

abstract_inverted_index.wetland

abstract_inverted_index.Cl

abstract_inverted_index.months,

abstract_inverted_index.rapidly

abstract_inverted_index.identification

abstract_inverted_index.hundreds

abstract_inverted_index.short

abstract_inverted_index.coming

abstract_inverted_index.always

abstract_inverted_index.demand,

abstract_inverted_index.cholera

abstract_inverted_index.Sanitation

abstract_inverted_index.successful

abstract_inverted_index.users,

abstract_inverted_index.Four

abstract_inverted_index.stakeholder

abstract_inverted_index.engage

abstract_inverted_index.future.